Usuario listo

+++ eduSoporteProv-main/eduSoporteProv/src/main/java/com/edutech/eduSoporteProv/service/UsuarioService.java
@@ -1,6 +1,7 @@
 package com.edutech.eduSoporteProv.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,22 +11,48 @@ import com.edutech.eduSoporteProv.repository.UsuarioRepository;
 
 @Service
 public class UsuarioService {
+    private final UsuarioRepository usuarioRepository;
+
     @Autowired
-    private UsuarioRepository usuarioRepository;
-    
-    public Usuario save(Usuario usuario) {
-        return usuarioRepository.save(usuario);
-    }
-    
-    public List<Usuario> findAll() {
-        return usuarioRepository.findAll();
+    public UsuarioService(UsuarioRepository usuarioRepository) {
+        this.usuarioRepository = usuarioRepository;
     }
 
-    public Usuario findxId(int id) {
+    public Usuario crearUsuario(Usuario usuario) {
+        return usuarioRepository.save(usuario);
+    }
+
+    public Optional<Usuario> buscarxId(int id) {
         return usuarioRepository.findById(id);
     }
 
-    public Usuario deletexId(int id) {
-        return usuarioRepository.deleteById(id);
+    public Optional<Usuario> actualizarUsuario(int id, Usuario updatedUsuario) {
+        return usuarioRepository.findById(id).map(usuario -> {
+            usuario.setCorreo(updatedUsuario.getCorreo());
+            usuario.setPassword(updatedUsuario.getPassword());
+            usuario.setUsuario(updatedUsuario.getUsuario());
+            usuario.setNombrereal(updatedUsuario.getNombrereal());
+            usuario.setRolUsuario(updatedUsuario.getRolUsuario());
+            return usuarioRepository.save(usuario);
+        });
+    }
+
+    public boolean eliminarUsuario(int id) {
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+
+    public boolean desactivarUsuario(int id){
+        Optional<Usuario> usuario = usuarioRepository.findById(id);
+        if(usuario.isPresent()) {
+            Usuario actUsuario = usuario.get();
+            actUsuario.setEsvisible(!actUsuario.isEsvisible());
+            usuarioRepository.save(actUsuario);
+            return true;
+        }
+        return false;
     }
 }
